change_detection_backend/requirements.txt
Flask
Flask-CORS
Pillow
numpy
torch
torchvision
python-dotenv